requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]